+++ syscall/samples/flock/flock-fork.c
@@ -8,29 +8,8 @@
 #include <sys/file.h>
 #include <sys/wait.h>
 
-/**
- * duplicate file descriptors (created by, for example, fork(2) or dup(2))
- * refer to the same lock. (see flock(2))
- */
-int open_and_write(const char *filename, const char *prefix, const char *msg)
-{
-	int ret;
-	int fd = open(filename, O_TRUNC | O_RDWR, 0644);
+#include "flock_helpers.h"
 
-	ret = flock(fd, LOCK_EX | LOCK_NB);
-	if (ret != 0) {
-		fprintf(stderr, "[%s] flock: %s\n", prefix, strerror(errno));
-		close(fd);
-		return -1;
-	}
-
-	write(fd, msg, strlen(msg));
-
-	flock(fd, LOCK_UN);
-	close(fd);
-
-	return 0;
-}
 
 int main(void)
 {
@@ -45,32 +24,25 @@ int main(void)
 	close(fd);
 
 	pid = fork();
-	/* Children */
-	if (pid == 0) {
-
-		int try_times = 10;
-
-		while (try_times-- &&
-		       !open_and_write(filename, "Child",
-				       "hello from child.\n")) ;
-
-		exit(0);
-
-		/* Father */
-	} else if (pid > 0) {
-
-		int try_times = 10;
-
-		while (try_times-- &&
-		       !open_and_write(filename, "Father",
-				       "hello from parent.\n")) ;
-
-		wait(NULL);
-
-	} else {
+	if (pid < 0) {
 		fprintf(stderr, "Failed fork\n");
 		exit(1);
 	}
 
+	/* Children */
+	if (pid == 0) {
+		int try_times = 10;
+		while (try_times-- &&
+		       !open_and_write_with_lock(filename, "Child",
+				   "hello from child.\n")) ;
+		exit(0);
+	}
+
+	/* Father */
+	int try_times = 10;
+	while (try_times-- &&
+	       !open_and_write_with_lock(filename, "Father",
+				  "hello from parent.\n")) ;
+	wait(NULL);
 	return 0;
 }