+++ syscall/samples/flock/flock_helpers.h
@@ -1,0 +1,4 @@
+#pragma once
+
+int open_and_write_with_lock(const char *filename, const char *prefix,
+			     const char *msg);