+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to
 - libvirt: Introduce from ostools.
 - top-Makefile: Apply sub-dir to user targets.
 #### Added
+- flock(2): add `flock_helpers.c`.
 - qemu: qcow2: Parse qcow2 file.
 - libvirt: add memory directory.
 - initrd: Add minimal image from ostools.